Sécurisation de l'espace d'administration en vérifiant le statut de l'utilisateur pour les actions du BackController.php. Il serait bien de factoriser les appels et rendre générique l'accès à certaines méthodes. Finalisation du projet sur le site whatzeweb. Améliorations à venir sur cette base peut être un jour

+++ src/controller/BackController.php
@@ -17,27 +17,29 @@ class BackController extends Controller
      */
     public function addArticle(Parameter $post)
     {
-        $post->trimAll();
-        //Si le formulaire d'ajout à été soumis
-        if ($post->get('submit')) {
-            //Validation des données avant soumission à la BD
-            $errors = $this->validation->validate($post, 'Article');
-            if (!$errors) {
-                $this->articleDAO->addArticle($post, $this->session->getUserInfo('id'));
-                //Création d'un message à afficher dans la session
-                $this->session->set('add_article', 'Le nouvel article à bien été ajouté');
-                //TODO: Redirection vers l'article crée : nécessite de récupérer son id après création
-                header('Location: ../public/index.php?route=administration');
+        if ($this->checkAdmin()){
+            $post->trimAll();
+            //Si le formulaire d'ajout à été soumis
+            if ($post->get('submit')) {
+                //Validation des données avant soumission à la BD
+                $errors = $this->validation->validate($post, 'Article');
+                if (!$errors) {
+                    $this->articleDAO->addArticle($post, $this->session->getUserInfo('id'));
+                    //Création d'un message à afficher dans la session
+                    $this->session->set('add_article', 'Le nouvel article à bien été ajouté');
+                    //TODO: Redirection vers l'article crée : nécessite de récupérer son id après création
+                    header('Location: ../public/index.php?route=administration');
+                } else {
+                    //Si il y a des erreurs, tjrs en modification avec données et errors en plus
+                    $this->view->render('add_article', [
+                        'post' => $post,
+                        'errors' => $errors
+                    ]);
+                }
             } else {
-                //Si il y a des erreurs, tjrs en modification avec données et errors en plus
-                $this->view->render('add_article', [
-                    'post' => $post,
-                    'errors' => $errors
-                ]);
+                //Si aucune données POST, création d'un article
+                $this->view->render('add_article');
             }
-        } else {
-            //Si aucune données POST, création d'un article
-            $this->view->render('add_article');
         }
     }
 
@@ -50,33 +52,35 @@ class BackController extends Controller
      */
     public function editArticle(Parameter $post, $articleId)
     {
-        $post->trimAll();
-        $article = $this->articleDAO->getArticle($articleId);
+        if ($this->checkAdmin()){
+            $post->trimAll();
+            $article = $this->articleDAO->getArticle($articleId);
 
-        if ($post->get('submit')) {
-            $errors = $this->validation->validate($post, 'Article');
-            if (!$errors) {
-                var_dump($this->session);
-                $this->articleDAO->editArticle($post, $articleId, $this->session->getUserInfo('id'));
-                $this->session->set('edit_article', 'L\'article à bien été mis à jour');
-                header('Location: ../public/index.php?route=administration');
+            if ($post->get('submit')) {
+                $errors = $this->validation->validate($post, 'Article');
+                if (!$errors) {
+                    var_dump($this->session);
+                    $this->articleDAO->editArticle($post, $articleId, $this->session->getUserInfo('id'));
+                    $this->session->set('edit_article', 'L\'article à bien été mis à jour');
+                    header('Location: ../public/index.php?route=administration');
+                } else {
+                    //Si il y a des erreurs, affichage avec les données soumises et erreurs
+                    $this->view->render('edit_article', [
+                        'post' => $post,
+                        'errors' => $errors
+                    ]);
+                }
             } else {
-                //Si il y a des erreurs, affichage avec les données soumises et erreurs
+                //Edition d'un article, données brutes de la base
+                $post->set('id', $article->getId());
+                $post->set('title', $article->getTitle());
+                $post->set('content', $article->getContent());
+                $post->set('author', $article->getAuthor());
+                //Si le formulaire n'a pas été soumis, on affiche l'article à modifier
                 $this->view->render('edit_article', [
-                    'post' => $post,
-                    'errors' => $errors
+                    'post' => $post
                 ]);
             }
-        } else {
-            //Edition d'un article, données brutes de la base
-            $post->set('id', $article->getId());
-            $post->set('title', $article->getTitle());
-            $post->set('content', $article->getContent());
-            $post->set('author', $article->getAuthor());
-            //Si le formulaire n'a pas été soumis, on affiche l'article à modifier
-            $this->view->render('edit_article', [
-                'post' => $post
-            ]);
         }
     }
 
@@ -86,10 +90,12 @@ class BackController extends Controller
      */
     public function deleteArticle($articleId)
     {
-        $this->articleDAO->deleteArticle($articleId);
-        //TODO : vérifier si suppression effective pour enregistrer le message dans la session
-        $this->session->set('delete_article', 'Article supprimé avec succès');
-        header('Location: ../public/index.php?route=administration');
+        if ($this->checkAdmin()) {
+            $this->articleDAO->deleteArticle($articleId);
+            //TODO : vérifier si suppression effective pour enregistrer le message dans la session
+            $this->session->set('delete_article', 'Article supprimé avec succès');
+            header('Location: ../public/index.php?route=administration');
+        }
     }
 
     /**
@@ -98,9 +104,11 @@ class BackController extends Controller
      */
     public function deleteComment($commentId)
     {
-        $this->commentDAO->deleteComment($commentId);
-        $this->session->set('delete_comment', 'Suppression du commentaire effectuée');
-        header('Location: ../public/index.php?route=administration');
+        if ($this->checkAdmin()) {
+            $this->commentDAO->deleteComment($commentId);
+            $this->session->set('delete_comment', 'Suppression du commentaire effectuée');
+            header('Location: ../public/index.php?route=administration');
+        }
     }
 
     /**
@@ -108,7 +116,9 @@ class BackController extends Controller
      */
     public function profile()
     {
-        $this->view->render('profile');
+        if ($this->checkLoggedIn()) {
+            $this->view->render('profile');
+        }
     }
 
     /*
@@ -117,21 +127,22 @@ class BackController extends Controller
      */
     public function updatePassword(Parameter $post)
     {
-        if ($post->get('submit')) {
-            $errors = $this->validation->validate($post, 'User');
-            if (!$errors) {
-                $this->userDAO->updatePassword($post, $this->session->get('user'));
-                $this->session->set('update_password', 'Le mot de passe à bien été mis à jour');
-                header('Location: ../public/index.php?route=profile');
+        if ($this->checkLoggedIn()) {
+            if ($post->get('submit')) {
+                $errors = $this->validation->validate($post, 'User');
+                if (!$errors) {
+                    $this->userDAO->updatePassword($post, $this->session->get('user'));
+                    $this->session->set('update_password', 'Le mot de passe à bien été mis à jour');
+                    header('Location: ../public/index.php?route=profile');
+                } else {
+                    $this->view->render('updatePassword', [
+                        'errors' => $errors
+                    ]);
+                }
             } else {
-                $this->view->render('updatePassword', [
-                    'errors' => $errors
-                ]);
+                $this->view->render('updatePassword');
             }
-        } else {
-            $this->view->render('updatePassword');
         }
-
     }
 
     /**
@@ -139,10 +150,12 @@ class BackController extends Controller
      */
     public function logout()
     {
-        $this->session->destroy();
-        $this->session->start();
-        $this->session->set('logout', 'Déconnexion réussie');
-        header('Location: ../public/index.php');
+        if ($this->checkLoggedIn()) {
+            $this->session->destroy();
+            $this->session->start();
+            $this->session->set('logout', 'Déconnexion réussie');
+            header('Location: ../public/index.php');
+        }
     }
 
     /**
@@ -150,11 +163,13 @@ class BackController extends Controller
      */
     public function deleteAccount()
     {
-        $this->userDAO->deleteAccount($this->session->get('user'));
-        $this->session->destroy();
-        $this->session->start();
-        $this->session->set('delete_account', 'Votre compte à bien été supprimé');
-        header('Location: ../public/index.php');
+        if ($this->checkLoggedIn()) {
+            $this->userDAO->deleteAccount($this->session->get('user'));
+            $this->session->destroy();
+            $this->session->start();
+            $this->session->set('delete_account', 'Votre compte à bien été supprimé');
+            header('Location: ../public/index.php');
+        }
     }
 
     /**
@@ -162,28 +177,56 @@ class BackController extends Controller
      */
     public function administration()
     {
-        $articles = $this->articleDAO->getArticles();
-        $flagComments = $this->commentDAO->getFlagComments();
-        $users = $this->userDAO->getUsers();
-        $this->view->render('administration', [
-            'articles' => $articles,
-            'comments' => $flagComments,
-            'users' => $users
-        ]);
+        if ($this->checkAdmin()) {
+            $articles = $this->articleDAO->getArticles();
+            $flagComments = $this->commentDAO->getFlagComments();
+            $users = $this->userDAO->getUsers();
+            $this->view->render('administration', [
+                'articles' => $articles,
+                'comments' => $flagComments,
+                'users' => $users
+            ]);
+        }
     }
 
     public function unflagComment($commentId)
     {
-        $this->commentDAO->unflagComment($commentId);
-        $this->session->set('unflag_comment', 'Le commentaire a bien été désignalé');
-        header('Location: ../public/index.php?route=administration');
+        if ($this->checkAdmin()) {
+            $this->commentDAO->unflagComment($commentId);
+            $this->session->set('unflag_comment', 'Le commentaire a bien été désignalé');
+            header('Location: ../public/index.php?route=administration');
+        }
     }
 
     public function deleteUser($userId)
     {
-        $this->userDAO->deleteUser($userId);
-        $this->session->set('delete_user', 'L\'utilisateur a bien été supprimé');
-        header('Location: ../public/index.php?route=administration');
+        if ($this->checkAdmin()) {
+            $this->userDAO->deleteUser($userId);
+            $this->session->set('delete_user', 'L\'utilisateur a bien été supprimé');
+            header('Location: ../public/index.php?route=administration');
+        }
     }
+
+    private function checkLoggedIn()
+    {
+        if(!$this->session->getUserInfo('pseudo')) {
+            $this->session->set('need_login', 'Vous devez vous connecter pour accéder à cette page');
+            header('Location: ../public/index.php?route=login');
+        } else {
+            return true;
+        }
+    }
+
+    private function checkAdmin()
+    {
+        $this->checkLoggedIn();
+        if(!($this->session->getUserInfo('role') === 'admin')) {
+            $this->session->set('not_admin', 'Vous n\'avez pas le droit d\'accéder à cette page');
+            header('Location: ../public/index.php?route=profile');
+        } else {
+            return true;
+        }
+    }
+
 
 }
